+++ PrimeFactors/PrimeFactorsTests/PrimeFactorsTests.swift
@@ -7,22 +7,7 @@
 //
 
 import XCTest
-
-class PrimeFactors {
-    
-    func generate(_ n: inout Int) -> [Int] {
-        var factors = [Int]()
-        var divisor = 2
-        while n > 1 {
-            while n % divisor == 0 {
-                factors.append(divisor)
-                n /= divisor
-            }
-            divisor += 1
-        }
-        return factors
-    }
-}
+import PrimeFactors
 
 class PrimeFactorsTests: XCTestCase {
